Initial:  initial thinking in Scheme.
Trying to think through a recursive-descent parser in Scheme
for simple arithmetic expressions, trying to capture elementary
arithmetic "MDAS" rules in therein, with infix notation, and
parentheses to override default precedenct.  Goal is to
parse expressions and derive a postorder parse tree traversal
which will yield the RPN version of the expression evaluation.