Further thoughts:  restructuring of parser is likely needed.
We likely need to rework the general expression parsing to thread the
list-of-remaining-tokens throughout calls to parsing procedures,
to reflect reality that we need to be able to make progress by "consuming"
tokens, and subsequent parsing operations need to be fed such lists.
This is less than ideal, since the "state-of-parsing" is thus explicitly
handed into each parsing procedure.