Sum and product:  first part of parsing sums and products.
Thus far, this only parses sums.  Tested this out first, to see
if the parsing in a single procedure could possibly work out.
The implementation of the product-parsing functions should be similar
to what was done for sums, with the wrinkle that the parsing
of the "factors" will leave open the possibility for recursive
invocation of sum-parsing, for "factors" which are actually
sums that are surrounded by bracket-characters (conventionally,
those are parentheses, but they will be represented as <LB>
and <RB> in the grammar, for "left bracket" and "right bracket.")