Refinements:  corrections to sum-parsing.
It helped to think about the mutually-recursive
parsing procedures in terms of structural induction
and the grammar defining the language.

For example, it's possible for all of the parsing
to be done by a recursive call and the results should
be examined in order to determine if that is what
happened.  The parsing function for sums does this
already, and the parsing function for products probably
should to this also.